\docType{data}
\name{areaDataFile}
\alias{areaDataFile}
\title{areaDataFile}
\source{
  Downloaded from the Canadian Open Data
  \href{http://www.data.gc.ca/default.asp?lang=En&n=5175A6F0-1&xsl=datacataloguerecord&metaxsl=datacataloguerecord&formid=C87D5FDD-00E6-41A0-B5BA-E8E41B521ED0}{database}
}
\description{
  areaDataFile
}
\details{
  Contains the latitudes (lat), longitudes (long), x (lamx)
  and y (lamy) point locations in the Lambert projection,
  the area id (uid), the number of dwellings (dwell), the
  total population (pop, based on the sum of the containing
  blocks), and the province of origin (prov). See the
  vignette for the processing performed to generate the
  file.
}
\keyword{datasets}

