\name{searchLoc}
\alias{searchLoc}
\title{takes an querystring, parses it, and returns the latitudes and longitudes}
\usage{
  searchLoc(inSearch)
}
\arguments{
  \item{inSearch}{the query string to use}
}
\description{
  takes an querystring, parses it, and returns the
  latitudes and longitudes
}

