\docType{data}
\name{timsLocs}
\alias{timsLocs}
\title{timsLocs}
\source{
  Foursquare, generated on June 1, 2013
}
\description{
  timsLocs
}
\details{
  Contains latitudes (lat), longitudes (long), and a
  textual description of Tim Hortons locations from
  Foursquare based on searching Foursquare for nearest
  locations to the lat and long from the areas in
  "areaData". This data will not be updated regularly.
}
\keyword{datasets}

